requirements.txt
numpy
cython
aubio
python-rtmidi
python-osc
PyQt5
PySide2
qtpy
Pillow
IPython
jupyter
glumpy
qtconsole
PyOpenGL
torch >= 1.4.0
torchvision
opencv-python
tensorboard >= 1.14